+++ arctic/static/arctic/src/assets/js/components/widgets.js
@@ -57,6 +57,11 @@ $(document).ready(function() {
                 if ($(instance).next().find('.item').length == 0) {
                     $(instance).next().next().removeAttr('style');
                 }
+            },
+            // Selectize reoders search results to find best match.
+            // But we wan't to keep original order by default
+            score: function() {
+              return function() { return 1 };
             }
         });
     });
@@ -83,6 +88,11 @@ $(document).ready(function() {
                     value: input,
                     text: input
                 }
+            },
+            // Selectize reoders search results to find best match.
+            // But we wan't to keep original order by default
+            score: function() {
+              return function() { return 1 };
             }
         });
     });
@@ -119,6 +129,11 @@ $(document).ready(function() {
                     $(instance).next().next().removeAttr('style');
                 }
             },
+            // Selectize reoders search results to find best match.
+            // But we wan't to keep original order by default
+            score: function() {
+              return function() { return 1 };
+            }
         });
     });