+++ example/articles/inlines.py
@@ -18,3 +18,4 @@ class ImagesInline(InlineFormSet):
     model = Image
     fields = "__all__"
     layout = {"blah": "bah"}
+    sorting_field = "order"