+++ CHANGELOG.md
@@ -15,10 +15,12 @@ Always reference the ticket number at the end of the issue description.
 - BREAKING! Fixed response in DeletionMixin after `delete` method is called [#334][334]
 - Added default `select_multiple` attr for `QuickFiltersSelectMultiple` widget [#340][340]
 - Updated Selectize version [#334][334]
+- Disabled options ordering by selectize in all Selectize widgets [#344][344]
 
 [334]: //github.com/sanoma/django-arctic/issues/334
 [340]: //github.com/sanoma/django-arctic/issues/340
 [342]: //github.com/sanoma/django-arctic/issues/342
+[344]: //github.com/sanoma/django-arctic/issues/344
 
 
 ## 1.3.6 (2019-01-22)