+++ CHANGELOG.md
@@ -24,9 +24,11 @@ Always reference the ticket number at the end of the issue description.
 
 - BREAKING! Fixed response in DeletionMixin after `delete` method is called [#334][334]
 - Added default `select_multiple` attr for `QuickFiltersSelectMultiple` widget [#340][340]
+- Updated Selectize version [#334][334]
 
 [334]: //github.com/sanoma/django-arctic/issues/334
 [340]: //github.com/sanoma/django-arctic/issues/340
+[342]: //github.com/sanoma/django-arctic/issues/342
 
 
 ## 1.3.6 (2019-01-22)