+++ CHANGELOG.md
@@ -10,6 +10,16 @@ Always reference the ticket number at the end of the issue description.
 
 ## pending
 
+### Added
+- Added the creation of new inline forms dynamically with javascript
+- Added the ability to sort (with drag n drop) inline forms
+[201]: https://github.com/sanoma/django-arctic/issues/201
+
+### Changed
+- Javascript widgets are not initialised with anonymous functions anymore
+- Separated javascript widgets into their own files (selectize, pickers, float labels, inlines) 
+- Added new size (extra large) to modals
+
 ### Fixed
 
 - BREAKING! Fixed response in DeletionMixin after `delete` method is called [#334][334]